requirements.txt
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.1
httplib2==0.22.0
idna==3.6
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.0.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-telegram-bot-calendar==1.0.5
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soundfile==0.12.1
telebot==0.0.5
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
